fix: v3.8.29 CI 编译修复 — MemoryTools onDelete Long + McpManager kotlinx import
1. MemoryTools onDelete lambda 由 suspend (Int)->Unit 改为 (Long)->Unit,
   修复 GenerationHandler 234/298 deleteMemory Int→Long 类型不匹配
2. McpManager 补齐 kotlinx.serialization.json 核心 import
   (JsonElement/JsonPrimitive/JsonArray/buildJsonArray/buildJsonObject/
   jsonPrimitive/contentOrNull)

版本不变: versionName 3.8.29

+++ app/src/main/java/me/rerere/rikkahub/data/ai/tools/MemoryTools.kt
@@ -23,7 +23,7 @@ fun buildMemoryTools(
     json: Json,
     onCreation: suspend (String) -> AssistantMemory,
     onUpdate: suspend (Long, String) -> AssistantMemory,
-    onDelete: suspend (Int) -> Unit
+    onDelete: suspend (Long) -> Unit
 ): List<Tool> = listOf(
     Tool(
         name = "memory_tool",
